requirements.txt
Babel>=0.9.6
pbr>=0.6,<1.0
SQLAlchemy>=0.7.8,<=0.8.99
alembic>=0.4.1
amqplib>=0.6.1
anyjson>=0.3.3
argparse
eventlet>=0.13.0
kombu>=2.4.8
lockfile>=0.8
lxml>=2.3
WebOb>=1.2.3
greenlet>=0.3.2
futures>=2.1.3
sqlalchemy-migrate>=0.8.2
netaddr>=0.7.6
paramiko>=1.9.0
iso8601>=0.1.8
python-neutronclient>=2.3.3,<3
python-glanceclient>=0.9.0
python-keystoneclient>=0.5.0
stevedore>=0.14
websockify>=0.5.1,<0.6
oslo.config>=1.2.0
oslo.rootwrap
pecan>=0.2.0
six>=1.4.1
jsonpatch>=1.1
WSME>=0.6
Jinja2
pyghmi>=0.5.8
